task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,22 +10,22 @@
  */
 int main(void)
 {
-	unsigned long long prev, next, ans;
+	unsigned long prev, next, ans;
 	int i;
 
 	i = 1;
 	prev = 1;
 	next = 2;
-	printf("%llu, %llu, ", prev, next);
+	printf("%lu, %lu, ", prev, next);
 	while (i <= 97)
 	{
 		ans = prev + next;
 		prev = next;
 		next = ans;
 		if (i == 97)
-			printf("%llu\n", ans);
+			printf("%lu\n", ans);
 		else
-			printf("%llu, ", ans);
+			printf("%lu, ", ans);
 		i++;
 	}
 	return (0);
